# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FinTrack AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /investment/advice:
    post:
      operationId: getInvestmentAdvice
      tags: [investment]
      summary: Get investment recommendations based on financial profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestmentAdviceBody"
      responses:
        "200":
          description: Investment recommendations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvestmentAdviceResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    InvestmentRecommendation:
      type: object
      properties:
        category:
          type: string
        percentage:
          type: number
        amount:
          type: number
        description:
          type: string
        riskLevel:
          type: string
      required:
        - category
        - percentage
        - amount
        - description
        - riskLevel
    InvestmentAdviceBody:
      type: object
      properties:
        monthlyIncome:
          type: number
        monthlySavings:
          type: number
        riskTolerance:
          type: string
        investmentGoal:
          type: string
        age:
          type: integer
      required:
        - monthlyIncome
        - monthlySavings
    InvestmentAdviceResponse:
      type: object
      properties:
        totalInvestable:
          type: number
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/InvestmentRecommendation"
        summary:
          type: string
        emergencyFundAdvice:
          type: string
        savingsRate:
          type: number
      required:
        - totalInvestable
        - recommendations
        - summary
        - emergencyFundAdvice
        - savingsRate
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      required:
        - error