# lib/api-zod/src/generated/types/investmentAdviceResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FinTrack AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InvestmentRecommendation } from "./investmentRecommendation";

export interface InvestmentAdviceResponse {
  totalInvestable: number;
  recommendations: InvestmentRecommendation[];
  summary: string;
  emergencyFundAdvice: string;
  savingsRate: number;
}