# lib/api-zod/src/generated/types/investmentRecommendation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FinTrack AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InvestmentRecommendation {
  category: string;
  percentage: number;
  amount: number;
  description: string;
  riskLevel: string;
}